requirements.txt
botcity-framework-core>=1.0.0,<2.0
botcity-framework-web>=0.8.0,<1.0
botcity-email-plugin
botcity-telegram-plugin
telebot
botcity-excel-plugin



botcity-maestro-sdk>=0.3.3,<1.0
# Uncomment to use botcity-utils
# botcity-utils>=0.1.2

# Uncomment to use botcity-crawler-plugin
# botcity-crawler-plugin

# Uncomment to use Pandas
# pandas
